Improve parameter sanitization; fix React bugs (#52)
- This commit greatly improves parameter sanitization, and also fixes some bugs in the React front-end (most notably broken "Delete Page" and Next/Prev functionality) that were discovered during testing.

+++ api/src/routes/api.ts
@@ -7,11 +7,13 @@ import Job from "../models/Job";
 
 const router = express.Router();
 import { setupPassport, requireToken } from "./auth";
+import fs = require("fs");
 setupPassport(router);
 
 function getJobFromRequest(req): Job {
-    // TODO: sanitize parameters!
-    return new Job(holdingArea() + req.params.category + "/" + req.params.job);
+    return fs.existsSync(holdingArea() + req.params.category + "/" + req.params.job)
+        ? new Job(holdingArea() + req.params.category + "/" + req.params.job)
+        : null;
 }
 
 function holdingArea(): string {
@@ -19,41 +21,72 @@ function holdingArea(): string {
     return holdingArea.endsWith("/") ? holdingArea : holdingArea + "/";
 }
 
+function sanitizeParameters(customRules = {}) {
+    return function (req, res, next) {
+        const defaultRule = /^[-.a-zA-Z0-9_]+$/;
+        for (const x in req.params) {
+            if (!req.params[x].match(customRules[x] ?? defaultRule)) {
+                return res.status(400).json({ error: "invalid: " + x });
+            }
+        }
+        next();
+    };
+}
+
 router.get("/", requireToken, function (req, res) {
     const categoryCollection = new CategoryCollection(holdingArea());
     res.json(categoryCollection.raw());
 });
 
-router.get("/:category", requireToken, function (req, res) {
-    // TODO
-    // Sanitize incoming parameters
-    // 404 error for non-existent catgeory (if holding area + category is not a directory)
+router.get("/:category", sanitizeParameters(), requireToken, function (req, res) {
+    if (!fs.existsSync(holdingArea() + req.params.category)) {
+        return res.status(404).json({ error: "Not Found" });
+    }
     const category = new Category(holdingArea() + req.params.category);
     res.json(category.raw());
 });
 
-router.get("/:category/:job", requireToken, function (req, res) {
-    res.json(getJobFromRequest(req).metadata.raw);
+router.get("/:category/:job", sanitizeParameters(), requireToken, function (req, res) {
+    const job = getJobFromRequest(req);
+    if (job == null) {
+        return res.status(404).json({ error: "Job not found" });
+    }
+    res.json(job.metadata.raw);
 });
 
-router.get("/:category/:job/status", requireToken, function (req, res) {
-    res.json(getJobFromRequest(req).metadata.status);
+router.get("/:category/:job/status", sanitizeParameters(), requireToken, function (req, res) {
+    const job = getJobFromRequest(req);
+    if (job == null) {
+        return res.status(404).json({ error: "Job not found" });
+    }
+    res.json(job.metadata.status);
 });
 
-router.put("/:category/:job/derivatives", requireToken, function (req, res) {
-    getJobFromRequest(req).makeDerivatives();
+router.put("/:category/:job/derivatives", sanitizeParameters(), requireToken, function (req, res) {
+    const job = getJobFromRequest(req);
+    if (job == null) {
+        return res.status(404).json({ error: "Job not found" });
+    }
+    job.makeDerivatives();
     res.json({ status: "ok" });
 });
 
-router.put("/:category/:job/ingest", requireToken, function (req, res) {
-    getJobFromRequest(req).ingest();
+router.put("/:category/:job/ingest", sanitizeParameters(), requireToken, function (req, res) {
+    const job = getJobFromRequest(req);
+    if (job == null) {
+        return res.status(404).json({ error: "Job not found" });
+    }
+    job.ingest();
     res.json({ status: "ok" });
 });
 
-router.put("/:category/:job", requireToken, function (req, res) {
+router.put("/:category/:job", sanitizeParameters(), requireToken, function (req, res) {
     const job = getJobFromRequest(req);
     const raw = req.body;
     try {
+        if (job == null) {
+            return res.status(404).json({ error: "Job not found" });
+        }
         job.metadata.raw = raw;
         const problems = job.metadata.status.file_problems as Record<string, Array<string>>;
         if (problems.added.length > 0 || problems.deleted.length > 0) {
@@ -66,9 +99,7 @@ router.put("/:category/:job", requireToken, function (req, res) {
     }
 });
 
-router.get("/:category/:job/:image/:size", requireToken, async function (req, res) {
-    //TODO
-    //Sanitize incoming parameters
+router.get("/:category/:job/:image/:size", sanitizeParameters(), requireToken, async function (req, res) {
     const legalSizes: Record<string, string> = {
         thumb: "THUMBNAIL",
         medium: "MEDIUM",
@@ -77,24 +108,34 @@ router.get("/:category/:job/:image/:size", requireToken, async function (req, re
     const image: string = req.params.image;
     const size: string = req.params.size;
     const job = getJobFromRequest(req);
+    if (job == null) {
+        return res.status(404).json({ error: "Job not found" });
+    }
     const deriv = await job.getImage(image).derivative(legalSizes[size] ?? "THUMBNAIL");
     res.sendFile(deriv);
 });
 
-router.delete("/:category/:job/:image/*"),
-    requireToken,
-    async function (req, res) {
-        //TODO
-        //Sanitize incoming parameters
-        const image: string = req.params.image;
-        const job = getJobFromRequest(req);
-        const imageObj = job.getImage(image);
-        if (imageObj !== null) {
-            imageObj.delete();
-            res.json({ status: "ok" });
-        } else {
-            res.status(404).json({ status: "image missing" });
+router.delete(
+    "/:category/:job/:image/*",
+    router.delete(
+        "/:category/:job/:image/*",
+        sanitizeParameters({ 0: /^\*$/ }),
+        requireToken,
+        async function (req, res) {
+            const image: string = req.params.image;
+            const job = getJobFromRequest(req);
+            if (job == null) {
+                return res.status(404).json({ error: "Job not found" });
+            }
+            const imageObj = job.getImage(image);
+            if (imageObj !== null) {
+                imageObj.delete();
+                res.json({ status: "ok" });
+            } else {
+                res.status(404).json({ status: "image missing" });
+            }
         }
-    };
+    )
+);
 
 module.exports = router;
